requirements-ocr.txt
beautifulsoup4
pycryptodome
requests
lxml

numpy
opencv_python
Pillow
PyYAML
tensorflow
